requirements.txt
Django==2.1.3
django-crispy_forms==1.7.2
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-js-asset==1.0.0